module_5/requirements.txt
psycopg[binary]
python-dotenv
Flask>=2.3,<4
huggingface_hub>=0.23.0
llama-cpp-python>=0.2.90,<0.3.0
sphinx
sphinx-rtd-theme
pylint
pydeps
beautifulsoup4
